requirements.txt
twine==1.10.0
wheel==0.30.0
Sphinx==1.7.1
django-grappelli==2.14.2
alabaster==0.7.12
aldryn-boilerplates==0.8.0
dj-database-url==0.5.0
Django==3.1.1
django-hitcount
django-appconf==1.0.4
django-autoslug==1.9.8
django-classy-tags==2.0.0
django-cleanup==5.1.0
django-contrib-comments==1.9.2
django-cors-headers==2.0.2
django-crispy-forms==1.9.2
django-disqus==0.5
django-filter==2.3.0
django-forms-bootstrap==3.1.0
django-formtools==2.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-mptt-admin==2.0.0
django-sekizai==2.0.0
django-six==1.0.4
django-storages==1.10.1
django-taggit==1.3.0
django-tinymce==3.0.2
Django-tinymce-filebrowser==0.2.1
django-tinymce4-lite==1.8.0
django-treebeard==4.3.1
django-utils-six==2.0
matplotlib==3.0.2
multipledispatch==0.6.0
mysqlclient==2.0.1
tblib==1.3.2
urllib3==1.24.1
gunicorn
oauthlib==3.1.0
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
Pillow==7.2.0
psutil==5.4.8
psycopg2==2.8.5
pycurl==7.43.0.2
virtualenv==20.0.31
importlib-metadata==1.7.0
importlib-resources==3.0.0
whitenoise